feat: separate register page from login component
Extract registration logic into a dedicated RegisterComponent, simplifying
LoginComponent to only handle login. Each component now has its own template
and styles with no shared mode-switching logic.
