+++ frontend/src/app/app.routes.ts
@@ -1,6 +1,7 @@
 import { Routes } from '@angular/router';
 
 import { LoginComponent } from './pages/login/login';
+import { RegisterComponent } from './pages/register/register';
 import { ItemsComponent } from './pages/items/items';
 import { MyItemsComponent } from './pages/my-items/my-items';
 import { ItemDetailComponent } from './pages/item-detail/item-detail';
@@ -8,8 +9,8 @@ import { ItemFormComponent } from './pages/item-form/item-form';
 import { authGuard } from './core/guards/auth.guard';
 
 export const routes: Routes = [
-  { path: 'login', component: LoginComponent, data: { mode: 'login' } },
-  { path: 'register', component: LoginComponent, data: { mode: 'register' } },
+  { path: 'login', component: LoginComponent },
+  { path: 'register', component: RegisterComponent },
   { path: 'items', component: ItemsComponent, canActivate: [authGuard] },
   { path: 'items/me', component: MyItemsComponent, canActivate: [authGuard] },
   { path: 'items/:id', component: ItemDetailComponent, canActivate: [authGuard] },
