+++ frontend/src/app/pages/register/register.css
@@ -1,0 +1,142 @@
+/* Auth page: centered panel + KBTU logo (scoped to LoginComponent) */
+
+:host {
+  --text: #111827;
+  --muted: #6b7280;
+  --border: #e5e7eb;
+  --primary: #2aa8d6;
+  --primary-hover: #1e92bd;
+  --danger: #b00020;
+  display: block;
+}
+
+.auth-shell {
+  min-height: 100vh;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 0;
+  color: var(--text);
+  background: #f3f4f6;
+}
+
+.auth-card {
+  width: min(460px, 100%);
+  margin: 0 auto;
+  background: #fff;
+  box-shadow: 0 25px 80px rgba(17, 24, 39, 0.12);
+  border-radius: 16px;
+  overflow: hidden;
+}
+
+/* Form panel */
+.auth-panel {
+  padding: 40px 40px 36px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  background: #ffffff;
+}
+
+.logo-badge {
+  align-self: flex-start;
+  padding: 12px 16px;
+  margin-bottom: 24px;
+  background: #ffffff;
+  border-radius: 14px;
+  box-shadow: 0 1px 3px rgba(17, 24, 39, 0.08);
+  border: 1px solid var(--border);
+}
+
+.logo-img {
+  display: block;
+  max-width: min(220px, 100%);
+  height: auto;
+}
+
+.auth-title {
+  margin: 0 0 20px;
+  font-size: 28px;
+  font-weight: 700;
+  letter-spacing: -0.02em;
+}
+
+.auth-form {
+  display: grid;
+  gap: 14px;
+  width: 100%;
+}
+
+.auth-field label {
+  display: block;
+  font-size: 13px;
+  color: var(--muted);
+  margin: 0 0 6px;
+}
+
+.auth-field input {
+  width: 100%;
+  padding: 12px 12px;
+  border: 1px solid var(--border);
+  border-radius: 10px;
+  outline: none;
+  font-size: 14px;
+  background: #fff;
+}
+
+.auth-field input:focus {
+  border-color: rgba(42, 168, 214, 0.6);
+  box-shadow: 0 0 0 4px rgba(42, 168, 214, 0.15);
+}
+
+.auth-actions {
+  margin-top: 6px;
+}
+
+.auth-button {
+  appearance: none;
+  border: 0;
+  border-radius: 10px;
+  padding: 12px 20px;
+  font-size: 14px;
+  font-weight: 600;
+  cursor: pointer;
+  color: #fff;
+  background: var(--primary);
+  width: 100%;
+}
+
+.auth-button:hover {
+  background: var(--primary-hover);
+}
+
+.auth-footer {
+  margin-top: 20px;
+  font-size: 14px;
+}
+
+.auth-footer a {
+  color: var(--primary);
+  text-decoration: none;
+  font-weight: 500;
+}
+
+.auth-footer a:hover {
+  text-decoration: underline;
+}
+
+.auth-error {
+  margin: 0 0 16px;
+  padding: 10px 12px;
+  border: 1px solid rgba(176, 0, 32, 0.25);
+  background: rgba(176, 0, 32, 0.06);
+  color: var(--danger);
+  border-radius: 10px;
+  font-size: 14px;
+}
+
+@media (max-width: 900px) {
+  .auth-panel {
+    padding: 28px 22px 32px;
+  }
+}
